Move message hooks to lib/messages
